+++ hyperkitty/lib/mailman.py
@@ -20,9 +20,7 @@
 # Author: Aurelien Bompard <abompard@fedoraproject.org>
 #
 
-from __future__ import absolute_import, unicode_literals
-
-from django.utils.six.moves.urllib.error import HTTPError
+from urllib.error import HTTPError
 
 from django_mailman3.lib.cache import cache
 from django_mailman3.lib.mailman import get_mailman_client
@@ -37,13 +35,13 @@ class ModeratedListException(Exception):
     pass
 
 
-def subscribe(list_address, user, email=None, display_name=None):
+def subscribe(list_id, user, email=None, display_name=None):
     if email is None:
         email = user.email
     if display_name is None:
         display_name = "%s %s" % (user.first_name, user.last_name)
     client = get_mailman_client()
-    rest_list = client.get_list(list_address)
+    rest_list = client.get_list(list_id)
     subscription_policy = rest_list.settings.get(
         "subscription_policy", "moderate")
     # Add a flag to return that would tell the user they have been subscribed
@@ -60,21 +58,29 @@ def subscribe(list_address, user, email=None, display_name=None):
                 " to it before posting.")
 
         # not subscribed yet, subscribe the user without email delivery
-        member = rest_list.subscribe(
-            email, display_name, pre_verified=True, pre_confirmed=True)
+        try:
+            member = rest_list.subscribe(
+                email, display_name, pre_verified=True, pre_confirmed=True)
+        except HTTPError as e:
+            if e.code == 409:
+                logger.info("Subscription for %s to %s is already pending",
+                            email, list_id)
+                return subscribed_now
+            else:
+                raise
         # The result can be a Member object or a dict if the subscription can't
         # be done directly, or if it's pending, or something else.
         # Broken API :-(
         if isinstance(member, dict):
             logger.info("Subscription for %s to %s is pending",
-                        email, list_address)
+                        email, list_id)
             return subscribed_now
         member.preferences["delivery_status"] = "by_user"
         member.preferences.save()
         subscribed_now = True
         cache.delete("User:%s:subscriptions" % user.id, version=2)
         logger.info("Subscribing %s to %s on first post",
-                    email, list_address)
+                    email, list_id)
 
     return subscribed_now