+++ hyperkitty/management/commands/mailman_sync.py
@@ -23,10 +23,6 @@
 Sync properties from Mailman into HyperKitty
 """
 
-from __future__ import absolute_import, print_function, unicode_literals
-
-from optparse import make_option
-
 from django.core.management.base import BaseCommand, CommandError
 from hyperkitty.lib.mailman import sync_with_mailman
 from hyperkitty.management.utils import setup_logging
@@ -34,11 +30,11 @@ from hyperkitty.management.utils import setup_logging
 
 class Command(BaseCommand):
     help = "Sync properties from Mailman into HyperKitty"
-    option_list = BaseCommand.option_list + (
-        make_option(
+
+    def add_arguments(self, parser):
+        parser.add_argument(
             '--overwrite', action='store_true', default=False,
-            help="overwrite existing Mailman IDs in HyperKitty's database"),
-        )
+            help="overwrite existing Mailman IDs in HyperKitty's database")
 
     def handle(self, *args, **options):
         options["verbosity"] = int(options.get("verbosity", "1"))