+++ hyperkitty/tests/models/test_mailinglist.py
@@ -20,16 +20,18 @@
 # Author: Aurelien Bompard <abompard@fedoraproject.org>
 #
 
-from __future__ import absolute_import, print_function, unicode_literals
+from datetime import datetime, date, timedelta
+from email.message import EmailMessage
+from random import shuffle
 
-from datetime import datetime
-from email.message import Message
-
+from django.contrib.auth.models import User
 from django.utils.timezone import utc
 from django_mailman3.tests.utils import FakeMMList
 
 from hyperkitty.lib.incoming import add_to_list
 from hyperkitty.models import MailingList, Thread, ArchivePolicy
+from hyperkitty.models.mailinglist import (
+    RecentThreads, TopThreads, PopularThreads, FirstDate)
 from hyperkitty.tests.utils import TestCase
 
 
@@ -78,10 +80,10 @@ class MailingListTestCase(TestCase):
         # the get_threads_between method should return all threads that have
         # been active between the two specified dates, including the threads
         # started in between those dates but updated later
-        msg1 = Message()
+        msg1 = EmailMessage()
         msg1["From"] = "sender1@example.com"
         msg1["Message-ID"] = "<msg1>"
-        msg1["Date"] = "2015-02-15 00:00:00 UTC"
+        msg1["Date"] = "15 Feb 2015 00:00:00 UTC"
         msg1.set_payload("message 1")
         add_to_list(self.ml.name, msg1)
         # The thread started in Feb, it should show up in the Feb threads but
@@ -107,17 +109,17 @@ class MailingListTestCase(TestCase):
         # the get_threads_between method should return all threads that have
         # been active between the two specified dates, including the threads
         # started in between those dates but updated later
-        msg1 = Message()
+        msg1 = EmailMessage()
         msg1["From"] = "sender1@example.com"
         msg1["Message-ID"] = "<msg1>"
-        msg1["Date"] = "2015-02-15 00:00:00 UTC"
+        msg1["Date"] = "15 Feb 2015 00:00:00 UTC"
         msg1.set_payload("message 1")
         add_to_list(self.ml.name, msg1)
-        msg2 = Message()
+        msg2 = EmailMessage()
         msg2["From"] = "sender2@example.com"
         msg2["Message-ID"] = "<msg2>"
         msg2["In-Reply-To"] = "<msg1>"
-        msg2["Date"] = "2015-03-15 00:00:00 UTC"
+        msg2["Date"] = "15 Mar 2015 00:00:00 UTC"
         msg2.set_payload("message 2")
         add_to_list(self.ml.name, msg2)
         # The thread started in Feb, was updated in March. It should show up in
@@ -138,17 +140,17 @@ class MailingListTestCase(TestCase):
         # the get_threads_between method should return all threads that have
         # been active between the two specified dates, including the threads
         # started in between those dates but updated later
-        msg1 = Message()
+        msg1 = EmailMessage()
         msg1["From"] = "sender1@example.com"
         msg1["Message-ID"] = "<msg1>"
-        msg1["Date"] = "2015-01-15 00:00:00 UTC"
+        msg1["Date"] = "15 Jan 2015 00:00:00 UTC"
         msg1.set_payload("message 1")
         add_to_list(self.ml.name, msg1)
-        msg2 = Message()
+        msg2 = EmailMessage()
         msg2["From"] = "sender2@example.com"
         msg2["Message-ID"] = "<msg2>"
         msg2["In-Reply-To"] = "<msg1>"
-        msg2["Date"] = "2015-03-15 00:00:00 UTC"
+        msg2["Date"] = "15 Mar 2015 00:00:00 UTC"
         msg2.set_payload("message 2")
         add_to_list(self.ml.name, msg2)
         # The thread started in Jan, was updated in March. It should show up in
@@ -169,3 +171,124 @@ class MailingListTestCase(TestCase):
             datetime(2015, 3, 31, 0, 0, 0, tzinfo=utc),
             )
         self.assertEqual(march_threads.count(), 1)
+
+
+class RecentThreadsTestCase(TestCase):
+
+    def setUp(self):
+        self.ml = MailingList.objects.create(name="list@example.com")
+        self.cached_value = RecentThreads(self.ml)
+
+    def test_order(self):
+        # The Thread instances returned by get_or_set() should be returned in
+        # the order of the list returned by get_value().
+        today = date.today()
+        ids = list(range(1, 21))
+        shuffle(ids)
+        # Add the emails in random order
+        for i in ids:
+            msg_date = today - timedelta(days=i)
+            msg = EmailMessage()
+            msg["From"] = "sender@example.com"
+            msg["Message-ID"] = "<msg%d>" % i
+            msg["Date"] = "%s 00:00:00 UTC" % msg_date.strftime("%d %b %Y")
+            msg.set_payload("message %d" % i)
+            add_to_list(self.ml.name, msg)
+        # The RecentThreads value should be reverse-sorted by date.
+        self.assertListEqual(
+            [t.starting_email.message_id for t in self.cached_value()],
+            ["msg%d" % i for i in range(1, 21)]
+            )
+
+
+class TopThreadsTestCase(TestCase):
+
+    def setUp(self):
+        self.ml = MailingList.objects.create(name="list@example.com")
+        self.cached_value = TopThreads(self.ml)
+
+    def test_order(self):
+        # The Thread instances returned by get_or_set() should be returned in
+        # the order of the list returned by get_value().
+        email_counts = list(range(1, 21))
+        # Create the threads in random order
+        shuffle(email_counts)
+        for email_count in email_counts:
+            msg = EmailMessage()
+            msg["From"] = "sender@example.com"
+            msg["Message-ID"] = "<msg%d>" % email_count
+            msg["Date"] = datetime.now().strftime("%d %b %Y %H:%M:%S %Z")
+            msg.set_payload("message %d" % email_count)
+            add_to_list(self.ml.name, msg)
+            # Add the replies
+            for email_num in range(email_count):
+                msg = EmailMessage()
+                msg["From"] = "sender@example.com"
+                msg["Message-ID"] = "<msg%d-%d>" % (email_count, email_num)
+                msg["In-Reply-To"] = "<msg%d>" % email_count
+                msg["Date"] = datetime.now().strftime("%d %b %Y %H:%M:%S %Z")
+                msg.set_payload("message %d-%d" % (email_count, email_num))
+                add_to_list(self.ml.name, msg)
+        # The TopThreads value should be reverse-sorted by number of emails.
+        self.assertListEqual(
+            [t.starting_email.message_id for t in self.cached_value()],
+            ["msg%d" % i for i in range(20, 0, -1)]
+            )
+
+
+class PopularThreadsTestCase(TestCase):
+
+    def setUp(self):
+        self.ml = MailingList.objects.create(name="list@example.com")
+        self.cached_value = PopularThreads(self.ml)
+
+    def test_order(self):
+        # The Thread instances returned by get_or_set() should be returned in
+        # the order of the list returned by get_value().
+        # Create users to vote
+        users = []
+        for uid in range(20):
+            users.append(User.objects.create(username="user%d" % uid))
+        # Create the threads to be voted on
+        votes_count = list(range(1, 21))
+        shuffle(votes_count)
+        # Add the emails in random order
+        for votes_num in votes_count:
+            msg = EmailMessage()
+            msg["From"] = "sender@example.com"
+            msg["Message-ID"] = "<msg%d>" % votes_num
+            msg["Date"] = datetime.now().strftime("%d %b %Y %H:%M:%S %Z")
+            msg.set_payload("message %d" % votes_num)
+            msg_id = add_to_list(self.ml.name, msg)
+            # Vote on the thread
+            thread = Thread.objects.get(thread_id=msg_id)
+            for uid in range(votes_num):
+                thread.starting_email.vote(1, users[uid])
+        # The PopularThreads value should be reverse-sorted by vote.
+        self.assertListEqual(
+            [t.starting_email.message_id for t in self.cached_value()],
+            ["msg%d" % i for i in range(20, 0, -1)]
+            )
+
+
+class FirstDateTestCase(TestCase):
+
+    def setUp(self):
+        self.ml = MailingList.objects.create(name="list@example.com")
+        self.cached_value = FirstDate(self.ml)
+
+    def test_no_email(self):
+        self.assertIsNone(self.cached_value())
+
+    def test_date(self):
+        # The date should be the date of the first email in the list
+        today = date.today()
+        for i in range(1, 21):
+            msg_date = today - timedelta(days=i)
+            msg = EmailMessage()
+            msg["From"] = "sender@example.com"
+            msg["Message-ID"] = "<msg%d>" % i
+            msg["Date"] = "%s 00:00:00 UTC" % msg_date.strftime("%d %b %Y")
+            msg.set_payload("message %d" % i)
+            add_to_list(self.ml.name, msg)
+        self.assertEqual(self.cached_value(), today - timedelta(days=20))