+++ hyperkitty/tests/views/test_message.py
@@ -21,16 +21,14 @@
 # Author: Aurelien Bompard <abompard@fedoraproject.org>
 #
 
-from __future__ import absolute_import, print_function, unicode_literals
-
 import json
 import uuid
-from email.message import Message
+from email.message import EmailMessage
 
 from allauth.account.models import EmailAddress
 from mock import Mock, patch
 from django.contrib.auth.models import User
-from django.core.urlresolvers import reverse
+from hyperkitty.utils import reverse
 from django.core import mail
 from django.utils import timezone
 from django_gravatar.helpers import get_gravatar_url
@@ -51,7 +49,7 @@ class MessageViewsTestCase(TestCase):
                 'testuser', 'test@example.com', 'testPass')
         self.client.login(username='testuser', password='testPass')
         # Create a dummy message to test on
-        msg = Message()
+        msg = EmailMessage()
         msg["From"] = "Dummy Sender <dummy@example.com>"
         msg["Subject"] = "Dummy Subject"
         msg["Date"] = "Mon, 02 Feb 2015 13:00:00 +0300"
@@ -64,7 +62,7 @@ class MessageViewsTestCase(TestCase):
                       get_message_id_hash("msg")))
         resp = self.client.post(url, {"vote": "1"})
         self.assertEqual(resp.status_code, 200)
-        result = json.loads(resp.content)
+        result = json.loads(resp.content.decode(resp.charset))
         self.assertEqual(result["like"], 1)
         self.assertEqual(result["dislike"], 0)
 
@@ -73,12 +71,12 @@ class MessageViewsTestCase(TestCase):
                       get_message_id_hash("msg")))
         resp = self.client.post(url, {"vote": "-1"})
         self.assertEqual(resp.status_code, 200)
-        result = json.loads(resp.content)
+        result = json.loads(resp.content.decode(resp.charset))
         self.assertEqual(result["like"], 0)
         self.assertEqual(result["dislike"], 1)
 
     def test_vote_cancel(self):
-        msg = Message()
+        msg = EmailMessage()
         msg["From"] = "dummy@example.com"
         msg["Message-ID"] = "<msg1>"
         msg.set_payload("Dummy message")
@@ -101,7 +99,7 @@ class MessageViewsTestCase(TestCase):
             votes = msg.get_votes()
             self.assertEqual(votes["likes"], 0)
             self.assertEqual(votes["dislikes"], 0)
-            result = json.loads(resp.content)
+            result = json.loads(resp.content.decode(resp.charset))
             self.assertEqual(result["like"], 0)
             self.assertEqual(result["dislike"], 0)
 
@@ -149,7 +147,7 @@ class MessageViewsTestCase(TestCase):
                 posting_fn.call_args[0][1:],
                 (mlist, 'Re: Dummy Subject', 'dummy reply content',
                  {'References': '<msg>', 'In-Reply-To': '<msg>'}))
-        result = json.loads(response.content)
+        result = json.loads(response.content.decode(response.charset))
         self.assertIn("Django User", result["message_html"])
         self.assertIn("dummy reply content", result["message_html"])
         self.assertIn(
@@ -170,7 +168,7 @@ class MessageViewsTestCase(TestCase):
             self.assertEqual(
                 posting_fn.call_args[0][1:],
                 (mlist, 'new subject', 'dummy reply content', {}))
-        result = json.loads(response.content)
+        result = json.loads(response.content.decode(response.charset))
         self.assertEqual(result["message_html"], None)
 
     def test_reply_different_sender(self):
@@ -200,7 +198,7 @@ class MessageViewsTestCase(TestCase):
                 (mlist, 'Re: Dummy Subject', 'dummy reply content',
                  {'From': 'otheremail@example.com',
                   'In-Reply-To': '<msg>', 'References': '<msg>'}))
-        result = json.loads(response.content)
+        result = json.loads(response.content.decode(response.charset))
         self.assertIn("Django User", result["message_html"])
         self.assertIn("dummy reply content", result["message_html"])
         self.assertIn(
@@ -282,7 +280,7 @@ class MessageViewsTestCase(TestCase):
         self.assertEqual(messages[0].tags, "success")
 
     def test_display_fixed(self):
-        msg = Message()
+        msg = EmailMessage()
         msg["From"] = "Dummy Sender <dummy@example.com>"
         msg["Subject"] = "Dummy Subject"
         msg["Date"] = "Mon, 02 Feb 2015 13:00:00 +0300"
@@ -301,7 +299,7 @@ class MessageViewsTestCase(TestCase):
         self.assertContains(response2, "email-body fixed", status_code=200)
 
     def test_email_escaped_body(self):
-        msg = Message()
+        msg = EmailMessage()
         msg["From"] = "Dummy Sender <dummy@example.com>"
         msg["Subject"] = "Dummy Subject"
         msg["Date"] = "Mon, 02 Feb 2015 13:00:00 +0300"
@@ -314,7 +312,7 @@ class MessageViewsTestCase(TestCase):
         self.assertNotContains(response, "email@example.com", status_code=200)
 
     def test_email_in_link_in_body(self):
-        msg = Message()
+        msg = EmailMessage()
         msg["From"] = "Dummy Sender <dummy@example.com>"
         msg["Subject"] = "Dummy Subject"
         msg["Date"] = "Mon, 02 Feb 2015 13:00:00 +0300"
@@ -330,7 +328,7 @@ class MessageViewsTestCase(TestCase):
             status_code=200)
 
     def test_email_escaped_sender(self):
-        msg = Message()
+        msg = EmailMessage()
         msg["From"] = "someone-else@example.com"
         msg["Subject"] = "Dummy Subject"
         msg["Date"] = "Mon, 02 Feb 2015 13:00:00 +0300"
@@ -373,7 +371,7 @@ class MessageViewsTestCase(TestCase):
         self.user.is_staff = True
         self.user.save()
         msg = Email.objects.get(message_id="msg")
-        msg2 = Message()
+        msg2 = EmailMessage()
         msg2["From"] = "dummy@example.com"
         msg2["Message-ID"] = "<msg2>"
         msg2["In-Reply-To"] = "<msg>"
@@ -406,7 +404,7 @@ class MessageViewsTestCase(TestCase):
         self.user.is_staff = True
         self.user.save()
         msg = Email.objects.get(message_id="msg")
-        msg2 = Message()
+        msg2 = EmailMessage()
         msg2["From"] = "dummy@example.com"
         msg2["Message-ID"] = "<msg2>"
         msg2["In-Reply-To"] = "<msg>"
@@ -434,7 +432,7 @@ class MessageViewsTestCase(TestCase):
         self.user.save()
         msg = Email.objects.get(message_id="msg")
         # Add a message with the same message-id to a different list.
-        msg2 = Message()
+        msg2 = EmailMessage()
         msg2["From"] = "dummy@example.com"
         msg2["Message-ID"] = "<msg>"
         msg2.set_payload("Dummy message")